created and wrote EmojiExtractor class
The class is written with clean architecture standards. All documentation included.

Added imageio and numpy to the requirements. Added a set of images to use for the time being. This is a very basic implementation.

Waiting for the rest of the team to decide on the next step, or make modifications if any changes are necessary.